fixed segfault when closing
git-svn-id: https://svn.osgeo.org/grass/grass/trunk@17428 15284696-431f-4ddb-bdfa-cd5b030d7da7

+++ vector/v.vol.rst/main.c
@@ -738,8 +738,10 @@ fprintf(stderr,"finished interpolating\n");
 	OUTGR ();
         if ((cellinp != NULL)) {
           G_close_cell(fdcout);
-	  fclose(Tmp_fd_cell);
-	  unlink(Tmp_file_cell);
+	  if ((cellout != NULL)) {
+		  fclose(Tmp_fd_cell);
+		  unlink(Tmp_file_cell);
+	  }
           G_close_cell(fdcell);
         }
         if (outz != NULL) {
